% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_contrataciones_estado.R
\name{scrap_contrataciones_estado}
\alias{scrap_contrataciones_estado}
\title{Scrapea la Plataforma de Contratación del Sector Público.}
\usage{
scrap_contrataciones_estado(palabra_clave)
}
\arguments{
\item{palabra_clave}{}
}
\value{
json
}
\description{
Scrapea la Plataforma de Contratación del Sector Público y filtra el resultado en función de la palabra clave parámetro, devolviendo un json
enviándo este a la plataforma Smart City (TECH) del cliente correspondiente.
}
\examples{
scrap_contrataciones_estado("suministro")

}
